+++ test/test_node_api.py
@@ -7,10 +7,9 @@ from unittest.mock import patch
 from urllib.parse import urlencode
 
 import pyinaturalist
-from pyinaturalist.constants import API_V1_BASE_URL, PER_PAGE_RESULTS
+from pyinaturalist.constants import API_V1_BASE_URL
 from pyinaturalist.exceptions import ObservationNotFound
 from pyinaturalist.node_api import (
-    get_all_observation_species_counts,
     get_all_observations,
     get_controlled_terms,
     get_geojson_observations,
@@ -172,11 +171,9 @@ def test_get_observations(requests_mock):
     assert len(first_result['place_ids']) == 13
 
 
-@patch('pyinaturalist.node_api.sleep')
-def test_get_all_observations(sleep, requests_mock):
-    # Make response appear as though there are more pages
+@patch('pyinaturalist.pagination.sleep')
+def test_get_observations__all_pages(sleep, requests_mock):
     page_1 = load_sample_data('get_observations_node_page1.json')
-    page_1['total_results'] = PER_PAGE_RESULTS + 1
     page_2 = load_sample_data('get_observations_node_page2.json')
 
     requests_mock.get(
@@ -187,10 +184,9 @@ def test_get_all_observations(sleep, requests_mock):
         ],
     )
 
-    observations = get_all_observations(id=[57754375, 57707611])
+    observations = get_observations(id=[57754375, 57707611], page='all')
 
-    assert type(observations) is list
-    assert len(observations) == 2
+    assert len(observations['results']) == 2
 
 
 def test_get_observation_observers(requests_mock):
@@ -278,7 +274,8 @@ def test_get_observation_species_counts(requests_mock):
     assert first_result['taxon']['name'] == 'Harmonia axyridis'
 
 
-def test_get_all_observation_species_counts(requests_mock):
+@patch('pyinaturalist.pagination.sleep')
+def test_get_observation_species_counts__all_pages(sleep, requests_mock):
     requests_mock.get(
         f'{API_V1_BASE_URL}/observations/species_counts',
         [
@@ -292,12 +289,13 @@ def test_get_all_observation_species_counts(requests_mock):
             },
         ],
     )
-    response = get_all_observation_species_counts(
-        user_login='my_username', quality_grade='research'
+    response = get_observation_species_counts(
+        user_login='my_username', quality_grade='research', page='all'
     )
-    first_result = response[0]
-    last_result = response[-1]
+    first_result = response['results'][0]
+    last_result = response['results'][-1]
 
+    assert len(response['results']) == 22
     assert first_result['count'] == 19
     assert first_result['taxon']['id'] == 27805
     assert first_result['taxon']['name'] == 'Notophthalmus viridescens'